requirements.txt
torch==2.2.1
filelock==3.13.1
typing-extensions==4.8.0
sympy==1.12
networkx==3.2.1
jinja2==3.1.2
fsspec==2024.3.0
mpmath==1.3.0
streamlit
Pillow
matplotlib
boto3
ultralytics
opencv-python